API: split the plans module up into parts
 - bluesky.plans : plans that produce run start / run stop
 - bluesky.plan_stubs : generators to yield from in complex plans
 - bluesky.preprocessors : functions which take in generators as input
 - bluesky.cntx : the deprecated plan context managers, likely to be removed
 - bluesky.object_plans : the state-ful object wrappers around plans

This helps to clarify to look for something based on it's function.
