+++ bluesky/__init__.py
@@ -7,25 +7,7 @@ from .utils import IllegalMessageSequence
 from .utils import FailedStatus
 
 from .run_engine import RunEngine
-from .plans import SupplementalData
-
-# for back-compat
-from .plans import PlanBase
-from .plans import Count
-from .plans import AbsListScanPlan
-from .plans import DeltaListScanPlan
-from .plans import AbsScanPlan
-from .plans import LogAbsScanPlan
-from .plans import DeltaScanPlan
-from .plans import LogDeltaScanPlan
-from .plans import AdaptiveAbsScanPlan
-from .plans import AdaptiveDeltaScanPlan
-from .plans import PlanND
-from .plans import InnerProductAbsScanPlan
-from .plans import InnerProductDeltaScanPlan
-from .plans import OuterProductAbsScanPlan
-from .plans import OuterProductDeltaScanPlan
-from .plans import Tweak
+from .preprocessors import SupplementalData
 
 from ._version import get_versions
 __version__ = get_versions()['version']
