+++ bluesky/tests/test_ramp_plan.py
@@ -1,6 +1,7 @@
 from bluesky.tests import requires_ophyd
 from bluesky.tests.utils import DocCollector
-from bluesky.plans import (ramp_plan, trigger_and_read)
+from bluesky.plans import ramp_plan
+from bluesky.plan_stubs import trigger_and_read
 from bluesky import Msg
 from bluesky.utils import RampFail
 import numpy as np
