+++ bluesky/tests/test_supplemental_data.py
@@ -1,6 +1,7 @@
 import pickle
 from bluesky import SupplementalData
-from bluesky.plans import count, pchain
+from bluesky.plans import count
+from bluesky.preprocessors import pchain
 
 
 def test_monitors(hw):
